+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to
 
 ### Added
 
+- Add payment methods for batch orders
 - Add unique reference for a quote once generated
 
 ## [3.0.4] - 2025-09-25